Create README.md
A phishing website is a common social engineering method that mimics trustful uniform resource locators (URLs) and webpages. The objective of this project is to train machine learning models on the dataset to predict phishing websites. Both phishing, Legitimate and suspicious URLs of websites are gathered to form a dataset and from them required URL and website content-based features are extracted. The performance level of each model is measured.